refactoring(Chess): refactoring styles and to segregate components
