+++ src/components/originalGames/Chess/components/Board/Board.jsx
@@ -2,13 +2,16 @@ import React from "react";
 import Square from "../Squares/Squares";
 import Queen from "../../pieces/Queen/Queen";
 import fillerPiece from "../../pieces/piece/Piece";
+import MatchInfo from "../MatchInfo/MatchInfo";
 import initializeBoard from "../../helpers/initializeBoard";
 import clearHighlight from "../../helpers/clearHighlight";
 import clearCheckHighlight from "../../helpers/clearCheckHighlight";
 import clearPossibleHighlight from "../../helpers/clearPossibleHighlight";
 import highlightMate from "../../helpers/highlightMate";
-import Notation from "../../helpers/notation";
 import calcSquareColor from "../../helpers/calcSquareColor";
+import invalidMove from "../../helpers/invalidMove";
+import makeMove from "../../helpers/makeMove";
+import { rowNums, colNums } from "../../constants/colsAndRows";
 import styles from "../../Game.module.css";
 
 export default class Board extends React.Component {
@@ -50,15 +53,7 @@ export default class Board extends React.Component {
   }
 
   reset() {
-    if (
-      this.state.historyNum - 1 === this.state.trueNum &&
-      this.state.turn === "1" &&
-      !this.state.mated
-    ) {
-      return "cannot reset";
-    }
-
-    this.setState({
+      this.setState({
       squares: initializeBoard(),
       source: -1,
       turn: "w",
@@ -97,17 +92,8 @@ export default class Board extends React.Component {
     let copySquares = squares.slice();
 
     copySquares = clearHighlight(copySquares).slice();
-    if (player === "w") {
-      copySquares = clearPossibleHighlight(copySquares).slice();
-      for (let j = 0; j < 64; j++) {
-        if (copySquares[j].ascii === "k") {
-          copySquares[j].inCheck = 0;
-          break;
-        }
-      }
-    }
+    copySquares = clearPossibleHighlight(copySquares).slice();
 
-    // note if king or rook has moved (castling not allowed if these have moved)
     if (copySquares[start].ascii === (player === "w" ? "k" : "K")) {
       if (player === "w") {
         this.setState({
@@ -143,7 +129,6 @@ export default class Board extends React.Component {
       }
     }
 
-    // add captured pieces to collection
     const collection =
       player === "w"
         ? this.state.piecesCollectedByWhite.slice()
@@ -154,10 +139,8 @@ export default class Board extends React.Component {
       });
     }
 
-    // make the move
-    copySquares = this.makeMove(copySquares, start, end).slice();
+    copySquares = makeMove(copySquares, start, end).slice();
 
-    // en passant helper
     const passantTrue =
       player === "w"
         ? copySquares[end].ascii === "p" &&
@@ -169,23 +152,14 @@ export default class Board extends React.Component {
           start <= 15 &&
           end - start === 16;
     let passant = passantTrue ? end : 65;
+    const turn = player === "w" ? "w" : "b";
 
-    // highlight mate
-    if (player === "w") {
-      copySquares = highlightMate(
-        "b",
-        copySquares,
-        this.checkmate("b", copySquares),
-        this.stalemate("b", copySquares)
-      ).slice();
-    } else {
-      copySquares = highlightMate(
-        "w",
-        copySquares,
-        this.checkmate("w", copySquares),
-        this.stalemate("w", copySquares)
-      ).slice();
-    }
+    copySquares = highlightMate(
+      turn,
+      copySquares,
+      this.checkmate(turn, copySquares),
+      this.stalemate(turn, copySquares)
+    ).slice();
 
     const copyHistory = this.state.history.slice();
     const copyHistoryH1 = this.state.historyH1.slice();
@@ -256,194 +230,6 @@ export default class Board extends React.Component {
     }
   }
 
-  // make a move
-  makeMove(squares, start, end, passantPos) {
-    const copySquares = squares.slice();
-    // castling
-    const isKing =
-      copySquares[start].ascii === "k" || copySquares[start].ascii === "K";
-    if (isKing && Math.abs(end - start) === 2) {
-      if (end === (copySquares[start].ascii === "k" ? 62 : 6)) {
-        copySquares[end - 1] = copySquares[end + 1];
-        copySquares[end - 1].highlight = 1;
-        copySquares[end + 1] = new fillerPiece(null);
-        copySquares[end + 1].highlight = 1;
-      } else if (end === (copySquares[start].ascii === "k" ? 58 : 2)) {
-        copySquares[end + 1] = copySquares[end - 2];
-        copySquares[end + 1].highlight = 1;
-        copySquares[end - 2] = new fillerPiece(null);
-        copySquares[end - 2].highlight = 1;
-      }
-    }
-
-    // en passant
-    const passant = passantPos === null ? this.state.passantPos : passantPos;
-    if (copySquares[start].ascii.toLowerCase() === "p") {
-      if (end - start === -7 || end - start === 9) {
-        if (start + 1 === passant)
-          copySquares[start + 1] = new fillerPiece(null);
-      } else if (end - start === -9 || end - start === 7) {
-        if (start - 1 === passant)
-          copySquares[start - 1] = new fillerPiece(null);
-      }
-    }
-
-    copySquares[end] = copySquares[start];
-    copySquares[end].highlight = 1;
-    copySquares[start] = new fillerPiece(null);
-    copySquares[start].highlight = 1;
-
-    // pawn promotion
-    if (copySquares[end].ascii === "p" && end >= 0 && end <= 7) {
-      copySquares[end] = new Queen("w");
-      copySquares[end].highlight = 1;
-    }
-    if (copySquares[end].ascii === "P" && end >= 56 && end <= 63) {
-      copySquares[end] = new Queen("b");
-      copySquares[end].highlight = 1;
-    }
-
-    return copySquares;
-  }
-
-  // returns true if castling is allowed
-  castlingAllowed(start, end, squares) {
-    const copySquares = squares.slice();
-    const player = copySquares[start].player;
-    const deltaPos = end - start;
-    if (start !== (player === "w" ? 60 : 4)) return false;
-    if (
-      (deltaPos === 2
-        ? copySquares[end + 1].ascii
-        : copySquares[end - 2].ascii) !== (player === "w" ? "r" : "R")
-    )
-      return false;
-    if (
-      (player === "w"
-        ? this.state.whiteKingHasMoved
-        : this.state.blackKingHasMoved) !== 0
-    )
-      return false;
-    if (player === "w") {
-      if (
-        (deltaPos === 2
-          ? this.state.rightWhiteRookHasMoved
-          : this.state.leftWhiteRookHasMoved) !== 0
-      )
-        return false;
-    } else if (player === "b") {
-      if (
-        (deltaPos === 2
-          ? this.state.rightBlackRookHasMoved
-          : this.state.leftBlackRookHasMoved) !== 0
-      )
-        return false;
-    }
-
-    return true;
-  }
-  blockersExist(start, end, squares) {
-    const startRow = 8 - Math.floor(start / 8);
-    const startCol = (start % 8) + 1;
-    const endRow = 8 - Math.floor(end / 8);
-    const endCol = (end % 8) + 1;
-    let rowDiff = endRow - startRow;
-    let colDiff = endCol - startCol;
-    let rowCtr = 0;
-    let colCtr = 0;
-    const copySquares = squares.slice();
-
-    while (colCtr !== colDiff || rowCtr !== rowDiff) {
-      let position = 64 - startRow * 8 + -8 * rowCtr + (startCol - 1 + colCtr);
-      if (
-        copySquares[position].ascii !== null &&
-        copySquares[position] !== copySquares[start]
-      )
-        return true;
-      if (colCtr !== colDiff) {
-        if (colDiff > 0) {
-          ++colCtr;
-        } else {
-          --colCtr;
-        }
-      }
-      if (rowCtr !== rowDiff) {
-        if (rowDiff > 0) {
-          ++rowCtr;
-        } else {
-          --rowCtr;
-        }
-      }
-    }
-    return false;
-  }
-  goodPawn(start, end, squares, passantPos) {
-    const passant = passantPos === null ? this.state.passantPos : passantPos;
-    const startRow = 8 - Math.floor(start / 8);
-    const startCol = (start % 8) + 1;
-    const endRow = 8 - Math.floor(end / 8);
-    const endCol = (end % 8) + 1;
-    const rowDiff = endRow - startRow;
-    const colDiff = endCol - startCol;
-    const copySquares = squares.slice();
-
-    if (rowDiff === 2 || rowDiff === -2) {
-      if (copySquares[start].player === "w" && (start < 48 || start > 55))
-        return false;
-      if (copySquares[start].player === "b" && (start < 8 || start > 15))
-        return false;
-    }
-    if (copySquares[end].ascii !== null) {
-      if (colDiff === 0) return false;
-    }
-    if (rowDiff === 1 && colDiff === 1) {
-      if (copySquares[end].ascii === null) {
-        if (copySquares[start + 1].ascii !== "P" || passant !== start + 1)
-          return false;
-      }
-    } else if (rowDiff === 1 && colDiff === -1) {
-      if (copySquares[end].ascii === null) {
-        if (copySquares[start - 1].ascii !== "P" || passant !== start - 1)
-          return false;
-      }
-    } else if (rowDiff === -1 && colDiff === 1) {
-      if (copySquares[end].ascii === null) {
-        if (copySquares[start + 1].ascii !== "p" || passant !== start + 1)
-          return false;
-      }
-    } else if (rowDiff === -1 && colDiff === -1) {
-      if (copySquares[end].ascii === null) {
-        if (copySquares[start - 1].ascii !== "p" || passant !== start - 1)
-          return false;
-      }
-    }
-
-    return true;
-  }
-  invalidMove(start, end, squares, passantPos) {
-    const copySquares = squares.slice();
-    console.log(copySquares);
-    const bqrpk =
-      copySquares[start].ascii.toLowerCase() === "r" ||
-      copySquares[start].ascii.toLowerCase() === "q" ||
-      copySquares[start].ascii.toLowerCase() === "b" ||
-      copySquares[start].ascii.toLowerCase() === "p" ||
-      copySquares[start].ascii.toLowerCase() === "k";
-    let invalid =
-      bqrpk === true && this.blockersExist(start, end, copySquares) === true;
-
-    if (invalid) return invalid;
-    const pawn = copySquares[start].ascii.toLowerCase() === "p";
-    invalid =
-      pawn === true &&
-      this.goodPawn(start, end, copySquares, passantPos) === false;
-    if (invalid) return invalid;
-    const king = copySquares[start].ascii.toLowerCase() === "k";
-    if (king && Math.abs(end - start) === 2)
-      invalid = this.castlingAllowed(start, end, copySquares) === false;
-
-    return invalid;
-  }
   canMoveThere(start, end, squares, passantPos) {
     const copySquares = squares.slice();
     if (start === end) return false;
@@ -454,7 +240,20 @@ export default class Board extends React.Component {
       copySquares[start].canMove(start, end) === false
     )
       return false;
-    if (this.invalidMove(start, end, copySquares, passantPos) === true)
+    if (
+      invalidMove(
+        start,
+        end,
+        copySquares,
+        passantPos,
+        this.state.whiteKingHasMoved,
+        this.state.blackKingHasMoved,
+        this.state.rightWhiteRookHasMoved,
+        this.state.leftWhiteRookHasMoved,
+        this.state.rightBlackRookHasMoved,
+        this.state.leftBlackRookHasMoved
+      ) === true
+    )
       return false;
 
     const cantCastle =
@@ -463,7 +262,6 @@ export default class Board extends React.Component {
       this.inCheck(player, copySquares);
     if (cantCastle) return false;
 
-    // king cannot castle through check
     if (
       copySquares[start].ascii === (player === "w" ? "k" : "K") &&
       Math.abs(end - start) === 2
@@ -475,7 +273,6 @@ export default class Board extends React.Component {
       if (this.inCheck(player, testSquares)) return false;
     }
 
-    // player cannot put or keep herself in check
     const checkSquares = squares.slice();
     checkSquares[end] = checkSquares[start];
     checkSquares[start] = new fillerPiece(null);
@@ -489,7 +286,6 @@ export default class Board extends React.Component {
     return true;
   }
 
-  // returns true if player is in check
   inCheck(player, squares) {
     let king = player === "w" ? "k" : "K";
     let positionOfKing = null;
@@ -505,13 +301,24 @@ export default class Board extends React.Component {
       if (copySquares[i].player !== player) {
         if (
           copySquares[i].canMove(i, positionOfKing) === true &&
-          this.invalidMove(i, positionOfKing, copySquares) === false
+          invalidMove(
+            i,
+            positionOfKing,
+            copySquares,
+            this.state.whiteKingHasMoved,
+            this.state.blackKingHasMoved,
+            this.state.rightWhiteRookHasMoved,
+            this.state.leftWhiteRookHasMoved,
+            this.state.rightBlackRookHasMoved,
+            this.state.leftBlackRookHasMoved
+          ) === false
         )
           return true;
       }
     }
     return false;
   }
+
   stalemate(player, squares) {
     if (this.inCheck(player, squares)) return false;
 
@@ -539,12 +346,6 @@ export default class Board extends React.Component {
 
   handleClick(i) {
     let copySquares = this.state.squares.slice();
-
-    if (this.state.historyNum - 1 !== this.state.trueNum) {
-      return "currently viewing history";
-    }
-
-    if (this.state.mated) return "game-over";
 
     if (this.state.source === -1) {
       if (copySquares[i].player !== this.state.turn) return -1;
@@ -621,52 +422,15 @@ export default class Board extends React.Component {
   }
 
   render() {
-    const rowNums = [];
-    for (let i = 8; i > 0; i--) {
-      rowNums.push(<Notation key={i} value={i} />);
-    }
-    const colNums = [];
-    for (let i = 1; i < 9; i++) {
-      let letter;
-      switch (i) {
-        case 1:
-          letter = "A";
-          break;
-        case 2:
-          letter = "B";
-          break;
-        case 3:
-          letter = "C";
-          break;
-        case 4:
-          letter = "D";
-          break;
-        case 5:
-          letter = "E";
-          break;
-        case 6:
-          letter = "F";
-          break;
-        case 7:
-          letter = "G";
-          break;
-        case 8:
-          letter = "H";
-          break;
-        default:
-          break;
-      }
-      colNums.push(<Notation key={letter} value={letter} />);
-    }
-
     const board = [];
     for (let i = 0; i < 8; i++) {
       const squareRows = [];
       for (let j = 0; j < 8; j++) {
         const copySquares = this.state.squares.slice();
         let squareColor = calcSquareColor(i, j, copySquares);
-        let squareCursor = "pointer";
-        if (copySquares[i * 8 + j].player !== "w") squareCursor = "default";
+        let squareCursor;
+        if (copySquares[i * 8 + j].player === this.state.turn)
+          squareCursor = "pointer";
 
         if (this.state.mated) squareCursor = "default";
         if (this.state.historyNum - 1 !== this.state.trueNum)
@@ -686,126 +450,38 @@ export default class Board extends React.Component {
     }
 
     return (
-      <div>
+      <div className={styles.game}>
         <div>
-          <div className={styles.left_screen}>
-            <div className={styles.side_box}>
-              <div className={styles.content + styles.title}>
-                <p className={styles.header_2_font}>Match Information</p>
-              </div>
-
-              <div className={styles.wrapper}>
-                <div className={styles.player_box}>
-                  <p className={styles.medium_font}>White</p>
-                </div>
-                <div className={styles.black_player_color}>
-                  <p className={styles.medium_font}>Black</p>
-                </div>
-              </div>
-              <div className={styles.wrapper}>
-                {this.state.turn === "w" ? (
-                  <div className={styles.highlight_box}></div>
-                ) : (
-                  <div className={styles.transparent}></div>
-                )}
-                {this.state.turn === "b" ? (
-                  <div className={styles.highlight_box}></div>
-                ) : (
-                  <div className={styles.transparent}></div>
-                )}
-              </div>
-
-              <div className={styles.button_wrapper}>
-                <button
-                  className={styles.reset_button}
-                  onClick={() => this.viewHistory("back_atw")}
-                >
-                  <p className={styles.button_font}>&lt;&lt;</p>
-                </button>
-                <button
-                  className={styles.reset_button}
-                  onClick={() => this.viewHistory("back")}
-                >
-                  <p className={styles.button_font}>&lt;</p>
-                </button>
-                <button
-                  className={styles.reset_button}
-                  onClick={() => this.reset()}
-                >
-                  <p className={styles.button_font}>Restart Game</p>
-                </button>
-                <button
-                  className={styles.reset_button}
-                  onClick={() => this.viewHistory("next")}
-                >
-                  <p className={styles.button_font}>&gt;</p>
-                </button>
-                <button
-                  className={styles.reset_button}
-                  onClick={() => this.viewHistory("next_atw")}
-                >
-                  <p className={styles.button_font}>&gt;&gt;</p>
-                </button>
-              </div>
-
-              <div className={styles.mate_wrapper}>
-                <p>
-                  {this.inCheck("w", this.state.squares) &&
-                  !this.checkmate("w", this.state.squares) === true
-                    ? "White is in check!"
-                    : ""}
-                </p>
-                <p>
-                  {this.inCheck("b", this.state.squares) &&
-                  !this.checkmate("b", this.state.squares) === true
-                    ? "Black is in check."
-                    : ""}
-                </p>
-                <p>
-                  {this.checkmate("w", this.state.squares) === true
-                    ? "Black won by checkmate."
-                    : ""}
-                </p>
-                <p>
-                  {this.checkmate("b", this.state.squares) === true
-                    ? "White won by checkmate!"
-                    : ""}
-                </p>
-                <p>
-                  {(this.stalemate("w", this.state.squares) &&
-                    this.state.turn === "w") === true
-                    ? "White are in stalemate. Game over."
-                    : ""}
-                </p>
-                <p>
-                  {(this.stalemate("b", this.state.squares) &&
-                    this.state.turn === "b") === true
-                    ? "Black is in stalemate. Game over."
-                    : ""}
-                </p>
-              </div>
+          <div className={styles.board}>
+            <div className={styles.row_label}> {rowNums} </div>
+            <div>
+              <div className={styles.table}> {board} </div>
+              <div className={styles.col_label}> {colNums} </div>
             </div>
           </div>
+          <MatchInfo
+            turn={this.state.turn}
+            backAtw={() => this.viewHistory("back_atw")}
+            back={() => this.viewHistory("back")}
+            reset={() => this.reset()}
+            next={() => this.viewHistory("next")}
+            nextAtw={() => this.viewHistory("next_atw")}
+          />
+        </div>
+        <div className={styles.wrapper}>
+          <div
+            className={this.state.turn === "w" ? styles.white_move : ""}
+          ></div>
 
-          <div className={styles.right_screen}>
-            <div className={styles.row_label}> {rowNums} </div>
-            <div className={styles.table}> {board} </div>
-            <div className={styles.col_label}> {colNums} </div>
-          </div>
+          <div
+            className={this.state.turn === "b" ? styles.black_move : ""}
+          ></div>
         </div>
       </div>
     );
   }
 
   viewHistory(direction) {
-    if (
-      this.state.historyNum - 1 === this.state.trueNum &&
-      this.state.turn === "1" &&
-      !this.state.mated
-    ) {
-      return "not allowed to view history";
-    }
-
     let copySquares = null;
     let copyWhiteCollection = null;
     let copyBlackCollection = null;
@@ -839,7 +515,7 @@ export default class Board extends React.Component {
       copyBlackCollection =
         this.state.historyBlackCollection[this.state.historyNum];
     } else {
-      return "no more history";
+      return null;
     }
 
     copySquares = clearPossibleHighlight(copySquares).slice();
