+++ src/components/originalGames/Chess/components/MatchInfo/MatchInfo.jsx
@@ -1,0 +1,33 @@
+import styles from "../../Game.module.css";
+
+export default function MatchInfo({
+  turn,
+  backAtw,
+  back,
+  reset,
+  next,
+  nextAtw,
+}) {
+  return (
+    
+      <div className={styles.side_box}>  
+        <div className={styles.button_wrapper}>
+          <button className={styles.button} onClick={backAtw}>
+            <p>&lt;&lt;</p>
+          </button>
+          <button className={styles.button} onClick={back}>
+            <p>&lt;</p>
+          </button>
+          <button className={styles.button} onClick={reset}>
+            <p>Restart</p>
+          </button>
+          <button className={styles.button} onClick={next}>
+            <p>&gt;</p>
+          </button>
+          <button className={styles.button} onClick={nextAtw}>
+            <p>&gt;&gt;</p>
+          </button>
+        </div>
+      </div>
+  );
+}
