+++ src/components/originalGames/Chess/helpers/highlightMate.js
@@ -1,14 +1,14 @@
 export default function highlightMate(
   player,
   squares,
-  check_mated,
-  stale_mated
+  checkMated,
+  staleMated
 ) {
   const copy_squares = squares.slice();
-  if (check_mated || stale_mated) {
+  if (checkMated || staleMated) {
     for (let j = 0; j < 64; j++) {
       if (copy_squares[j].ascii === (player === "w" ? "k" : "K")) {
-        copy_squares[j].checked = check_mated === true ? 1 : 2;
+        copy_squares[j].checked = checkMated === true ? 1 : 2;
         break;
       }
     }
